Extract GitLabClient from endpoints.gitlab.
